Fix scene load breaking when aspect is not present

+++ src/MeidoPhotoStudio.Plugin/Core/SceneManagement/LightAspectLoader.cs
@@ -18,6 +18,9 @@ public class LightAspectLoader(LightService lightService, BackgroundService back
         if (!loadOptions["lights"].Enabled)
             return;
 
+        if (lightsSchema is null)
+            return;
+
         lightService.RemoveAllLights();
 
         lightService.AddedLight += OnLightAdded;
